# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NASA API",
        "description": "API documentation for NASA API integration",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/nasa/earth-imagery": {
            "get": {
                "tags": [
                    "NASA Endpoints"
                ],
                "summary": "Get Earth Imagery",
                "operationId": "7d6067ce121e83da176b06fa2f5fd59e",
                "parameters": [
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Latitude",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "lon",
                        "in": "query",
                        "description": "Longitude",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Date",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "url": {
                                            "type": "string"
                                        },
                                        "id": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}